collector/lmstat_feature_exp: consolidate windows and linux implementations
